+++ paddle/phi/kernels/gpu/cross_entropy_grad_kernel.cu
@@ -99,7 +99,9 @@ __global__ void SoftCrossEntropyGradientKernel(T* logit_grad,
                                                const int64_t n,
                                                const int64_t d,
                                                const int64_t remain) {
-  int64_t ids = blockIdx.x * blockDim.x + threadIdx.x;
+  int64_t ids =
+      static_cast<int64_t>(blockIdx.x) * static_cast<int64_t>(blockDim.x) +
+      static_cast<int64_t>(threadIdx.x);
   if (ids < n * d) {
     int64_t idx_n = ids / d;
     int64_t idx_remain = ids % remain;
@@ -120,7 +122,9 @@ __global__ void SoftmaxWithCrossEntropyGradHardLabel(T* logits_grad,
                                                      const int64_t dim,
                                                      const int64_t d,
                                                      const int ignore_index) {
-  int64_t idx = blockIdx.x * blockDim.x + threadIdx.x;
+  int64_t idx =
+      static_cast<int64_t>(blockIdx.x) * static_cast<int64_t>(blockDim.x) +
+      static_cast<int64_t>(threadIdx.x);
   int64_t idx_n = idx / (d * dim);
   int64_t idx_dim = (idx / d) % dim;
   int64_t idx_d = idx % d;
